requirements.txt
docopt==0.6.2
arrow==0.17.0
Flask==1.1.2
flask_cors==3.0.9
flask_migrate==2.5.3
flask_restful==0.3.8
flask_sqlalchemy==2.4.4
flask_swagger_ui==3.36.0
pytest==6.1.2
python-dotenv==0.15.0
python_json_logger==2.0.1
